requirements.txt
bittensor
python-dotenv
requests
gspread>=6.0.0
google-auth>=2.0.0
